// packages/datadog-api-client-v1/models/SyntheticsBasicAuth.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Object to handle basic authentication when performing the test.
*/
export declare class SyntheticsBasicAuth {
    /**
    * Password to use for the basic authentication.
    */
    'password': string;
    /**
    * Username to use for the basic authentication.
    */
    'username': string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsBasicAuth;
    static serialize(data: SyntheticsBasicAuth): {
        [key: string]: any;
    };
    constructor();
}