// packages/datadog-api-client-v1/models/SyntheticsTestOptionsRetry.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Object describing the retry strategy to apply to a Synthetic test.
*/
export declare class SyntheticsTestOptionsRetry {
    /**
    * Number of times a test needs to be retried before marking a location as failed. Defaults to 0.
    */
    'count'?: number;
    /**
    * Time interval between retries (in milliseconds). Defaults to 300ms.
    */
    'interval'?: number;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsTestOptionsRetry;
    static serialize(data: SyntheticsTestOptionsRetry): {
        [key: string]: any;
    };
    constructor();
}